+++ modules-namespaces/src/components/base-components.ts
@@ -1,0 +1,36 @@
+// Component Base Class
+namespace App {
+  export abstract class Component<T extends HTMLElement, U extends HTMLElement> {
+    templateElement: HTMLTemplateElement;
+    hostElement: T;
+    element: U;
+
+    constructor(
+      templateId: string,
+      hostElementId: string,
+      insertAtStart: boolean,
+      newElementId?: string
+    ) {
+      this.templateElement = document.getElementById(templateId)! as HTMLTemplateElement;
+      this.hostElement = document.getElementById(hostElementId)! as T;
+
+      const importNode = document.importNode(this.templateElement.content, true);
+      this.element = importNode.firstElementChild as U;
+      if (newElementId) {
+        this.element.id = newElementId;
+      }
+
+      this.attach(insertAtStart);
+    }
+
+    private attach(insertAtBeggining: boolean) {
+      this.hostElement.insertAdjacentElement(
+        insertAtBeggining ? 'afterbegin' : 'beforeend',
+        this.element
+      );
+    }
+
+    abstract configure(): void;
+    abstract renderContent(): void;
+  }
+}
