+++ modules-namespaces/src/decorators/autobind.ts
@@ -1,0 +1,18 @@
+// autobind decorator
+namespace App {
+  export function autobind(
+    _: any,
+    _2: string,
+    descriptor: PropertyDescriptor
+  ) {
+    const originalMethod = descriptor.value;
+    const adjustedDescriptor: PropertyDescriptor = {
+      configurable: true,
+      get() {
+        const boundFunction = originalMethod.bind(this);
+        return boundFunction;
+      }
+    };
+    return adjustedDescriptor;
+  }
+}
